# backend/cmd/server/main.go
package main

import (
	"fmt"
	"log"
	"net/http"
	"os"
	"os/exec"
	"path/filepath"
	"strings"
	"time"

	"gohotel/internal/config"
	"gohotel/internal/database"
	"gohotel/internal/handler"
	"gohotel/internal/middleware"
	"gohotel/internal/repository"
	"gohotel/internal/service"
	"gohotel/pkg/logger"
	"gohotel/pkg/utils"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	_ "gohotel/docs" // 导入生成的 docs 包
)

// @title           酒店管理系统 API
// @version         1.0
// @description     这是一个酒店预订管理系统的 RESTful API 服务
// @termsOfService  http://swagger.io/terms/

// @contact.name   API Support
// @contact.email  support@gohotel.com

// @license.name  MIT
// @license.url   https://opensource.org/licenses/MIT

// @host      127.0.0.1:18080
// @BasePath

// @securityDefinitions.apikey Bearer
// @in header
// @name Authorization
// @description 输入 "Bearer {token}" 格式的 JWT token

func main() {
	// 1. 加载配置
	fmt.Println("📖 正在加载配置...")
	if err := config.Load(); err != nil {
		log.Fatal("配置加载失败:", err)
	}
	fmt.Println("✅ 配置加载成功!")

	backendDir := resolveBackendDir()
	frontendDistDir := resolveFrontendDistDir(backendDir)
	swaggerJSONPath := filepath.Join(backendDir, "docs", "swagger.json")
	if shouldAutoUpdateSwagger(config.AppConfig.Server.Mode) {
		if err := runSwagInit(backendDir); err != nil {
			log.Printf("⚠️  Swagger 文档自动更新失败: %v", err)
		} else {
			fmt.Println("✅ Swagger 文档已自动更新!")
		}
	}

	// 2. 初始化日志
	fmt.Println("📝 正在初始化日志...")
	if err := logger.Init(&logger.LogConfig{
		Level:      config.AppConfig.Log.Level,
		Filename:   config.AppConfig.Log.Filename,
		MaxSize:    config.AppConfig.Log.MaxSize,
		MaxBackups: config.AppConfig.Log.MaxBackups,
		MaxAge:     config.AppConfig.Log.MaxAge,
		Compress:   config.AppConfig.Log.Compress,
		Console:    config.AppConfig.Log.Console,
	}); err != nil {
		log.Fatal("日志初始化失败:", err)
	}
	defer logger.Sync()
	fmt.Println("✅ 日志初始化成功!")

	// 3. 连接数据库
	fmt.Println("🔌 正在连接数据库...")
	if err := database.InitMySQL(); err != nil {
		log.Fatal("数据库连接失败:", err)
	}
	defer database.CloseDB()

	// 4. 自动迁移数据库表
	fmt.Println("🔄 正在执行数据库迁移...")
	if err := database.AutoMigrate(); err != nil {
		log.Fatal("数据库迁移失败:", err)
	}

	// 5. 初始化雪花算法节点
	fmt.Println("❄️  正在初始化雪花算法节点...")
	// 节点ID可以从配置文件读取，这里暂时使用固定值 1
	// 如果是分布式部署，需要确保每个实例使用不同的节点ID（0-1023）
	if err := utils.InitSnowflake(1); err != nil {
		log.Fatal("雪花算法初始化失败:", err)
	}
	fmt.Println("✅ 雪花算法初始化成功!")

	// 6. 插入测试数据（可选）
	if err := database.SeedData(); err != nil {
		log.Fatal("测试数据插入失败:", err)
	}

	// 6.1 初始化COS服务
	fmt.Println("☁️  正在初始化COS服务...")
	var cosService *service.CosService
	var cosErr error
	cosService, cosErr = service.NewCosService(&config.AppConfig.COS)
	if cosErr != nil {
		log.Printf("⚠️  COS服务初始化失败: %v，将无法使用图片上传功能", cosErr)
		cosService = nil
	} else {
		fmt.Println("✅ COS服务初始化成功!")
	}

	// 6.2 初始化时间轮
	fmt.Println("⏰ 正在初始化时间轮...")
	timeWheel := utils.NewMultiTimeWheel() // 使用多层时间轮（秒、分、时、天四层）

	// 设置持久化存储，将任务保存到data目录
	persistStore := utils.NewFilePersistStore("./data/timewheel_tasks.json")
	timeWheel.SetPersistStore(persistStore)

	timeWheel.Start()
	defer timeWheel.Stop()
	fmt.Println("✅ 时间轮初始化成功!")

	// 7. 初始化依赖注入
	// Repository 层
	userRepo := repository.NewUserRepository(database.DB)
	hotelRepo := repository.NewHotelRepository(database.DB)
	hotelSettingsRepo := repository.NewHotelSettingsRepository(database.DB)
	roomRepo := repository.NewRoomRepository(database.DB)
	roomCategoryRepo := repository.NewRoomCategoryRepository(database.DB)
	bookingRepo := repository.NewBookingRepository(database.DB)
	logRepo := repository.NewLogRepository(database.DB)
	facilityRepo := repository.NewFacilityRepository(database.DB)
	bannerRepo := repository.NewBannerRepository(database.DB)
	noticeRepo := repository.NewNoticeRepository(database.DB)
	inventoryRepo := repository.NewInventoryRepository(database.DB)
	workOrderRepo := repository.NewWorkOrderRepository(database.DB)
	pricingRepo := repository.NewPricingRepository(database.DB)

	// Service 层
	userService := service.NewUserService(userRepo)
	hotelService := service.NewHotelService(database.DB, hotelRepo, hotelSettingsRepo)
	hotelSettingsService := service.NewHotelSettingsService(database.DB, hotelSettingsRepo, hotelRepo)
	noticeService := service.NewNoticeService(noticeRepo, cosService, timeWheel)
	inventoryService := service.NewInventoryService(inventoryRepo, pricingRepo)
	workOrderService := service.NewWorkOrderService(workOrderRepo, roomRepo)
	roomService := service.NewRoomService(roomRepo, roomCategoryRepo)
	roomCategoryService := service.NewRoomCategoryService(database.DB, roomCategoryRepo)
	bookingService := service.NewBookingService(bookingRepo, roomRepo, userRepo, userService, inventoryService)
	logService := service.NewLogService(logRepo)
	facilityService := service.NewFacilityService(facilityRepo)
	bannerService := service.NewBannerService(bannerRepo, cosService, timeWheel)
	recommendationService := service.NewRecommendationService(database.DB, roomRepo, userRepo)

	// 加载持久化的时间轮任务
	fmt.Println("📂 正在加载时间轮任务...")
	if err := timeWheel.LoadTasks(); err != nil {
		log.Printf("⚠️  时间轮任务加载失败: %v", err)
	} else {
		fmt.Println("✅ 时间轮任务加载成功!")
	}

	// 添加COS临时文件清理任务
	if cosService != nil {
		// 立即执行一次清理
		go func() {
			_, err := cosService.CleanupTempFiles(1 * time.Hour) // 清理1小时前的临时文件
			if err != nil {
				log.Printf("⚠️  初始清理COS临时文件失败: %v", err)
			}
		}()

		// 使用命名函数来实现递归调用
		var cleanupTask func()
		cleanupTask = func() {
			_, err := cosService.CleanupTempFiles(1 * time.Hour) // 清理1小时前的临时文件
			if err != nil {
				log.Printf("⚠️  清理COS临时文件失败: %v", err)
			}

			// 任务执行完成后，添加下一次任务
			nextExecTime := time.Now().Add(30 * time.Minute)
			timeWheel.AddTask(nextExecTime, cleanupTask, nil, true) // 不持久化任务
		}

		// 计算下一次执行时间
		nextExecTime := time.Now().Add(30 * time.Minute)

		// 添加任务到时间轮
		timeWheel.AddTask(nextExecTime, cleanupTask, nil, true) // 不持久化任务

		fmt.Println("✅ COS临时文件清理任务已添加，每30分钟执行一次")
	}

	// Handler 层
	userHandler := handler.NewUserHandler(userService)
	hotelHandler := handler.NewHotelHandler(hotelService)
	hotelSettingsHandler := handler.NewHotelSettingsHandler(hotelSettingsService)
	roomHandler := handler.NewRoomHandler(roomService)
	roomCategoryHandler := handler.NewRoomCategoryHandler(roomCategoryService)
	bookingHandler := handler.NewBookingHandler(bookingService)
	logHandler := handler.NewLogHandler(logService)
	facilityHandler := handler.NewFacilityHandler(facilityService)
	bannerHandler := handler.NewBannerHandler(bannerService, cosService)
	noticeHandler := handler.NewNoticeHandler(noticeService)
	cosHandler := handler.NewCosHandler(cosService)
	workOrderHandler := handler.NewWorkOrderHandler(workOrderService)
	pricingHandler := handler.NewPricingHandler(pricingRepo)
	inventoryHandler := handler.NewInventoryHandler(inventoryService)
	recommendationHandler := handler.NewRecommendationHandler(recommendationService)

	// 8. 设置 Gin 模式
	gin.SetMode(config.AppConfig.Server.Mode)

	// 9. 创建 Gin 引擎
	r := gin.New()

	// 10. 使用中间件
	r.Use(gin.Recovery())                // 恢复中间件（处理 panic）
	r.Use(middleware.CORSMiddleware())   // 跨域中间件
	r.Use(middleware.LoggerMiddleware()) // 日志中间件

	// 设置路由
	setupRoutes(r, userHandler, hotelHandler, hotelSettingsHandler, roomHandler, roomCategoryHandler, bookingHandler, logHandler, facilityHandler, bannerHandler, noticeHandler, cosHandler, workOrderHandler, pricingHandler, inventoryHandler, recommendationHandler, swaggerJSONPath, frontendDistDir)

	// 12. 启动服务器
	listenAddr := config.AppConfig.Server.ListenAddress()
	baseURL := config.AppConfig.Server.HTTPBaseURL()

	fmt.Println("═══════════════════════════════════════════════")
	fmt.Println("🏨 酒店管理系统 API 服务器")
	fmt.Println("═══════════════════════════════════════════════")
	fmt.Printf("📍 服务器地址: %s\n", baseURL)
	fmt.Printf("📝 运行模式: %s\n", config.AppConfig.Server.Mode)
	fmt.Printf("📚 Swagger 文档: %s/swagger/index.html\n", baseURL)
	fmt.Println("═══════════════════════════════════════════════")

	if err := r.Run(listenAddr); err != nil {
		log.Fatal("服务器启动失败:", err)
	}
}

// setupRoutes 设置所有路由

func setupRoutes(r *gin.Engine, userHandler *handler.UserHandler, hotelHandler *handler.HotelHandler, hotelSettingsHandler *handler.HotelSettingsHandler, roomHandler *handler.RoomHandler, roomCategoryHandler *handler.RoomCategoryHandler, bookingHandler *handler.BookingHandler, logHandler *handler.LogHandler, facilityHandler *handler.FacilityHandler, bannerHandler *handler.BannerHandler, noticeHandler *handler.NoticeHandler, cosHandler *handler.CosHandler, workOrderHandler *handler.WorkOrderHandler, pricingHandler *handler.PricingHandler, inventoryHandler *handler.InventoryHandler, recommendationHandler *handler.RecommendationHandler, swaggerJSONPath string, frontendDistDir string) {
	// Swagger 文档路由
	r.GET("/swagger.json", func(c *gin.Context) {
		c.File(swaggerJSONPath)
	})
	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler, ginSwagger.URL("/swagger.json")))

	// 健康检查
	// @Summary 健康检查
	// @Description 检查服务器运行状态
	// @Tags 系统
	// @Accept json
	// @Produce json
	// @Success 200 {object} map[string]interface{}
	// @Router /health [get]
	r.GET("/health", func(c *gin.Context) {
		c.JSON(200, gin.H{
			"status":  "ok",
			"message": "酒店管理系统运行正常",
		})
	})

	// API 路由组
	api := r.Group("/api")
	{
		// 认证路由（公开）
		auth := api.Group("/auth")
		{
			auth.POST("/register", userHandler.Register)
			auth.POST("/login", userHandler.Login)
			auth.POST("/wechat-login", userHandler.WeChatLogin)
		}

		// 房间路由（公开查询）
		rooms := api.Group("/rooms")
		{
			rooms.GET("", roomHandler.ListRooms)                     // 获取所有房间
			rooms.GET("/available", roomHandler.ListAvailableRooms)  // 获取可用房间
			rooms.GET("/floor/:floor", roomHandler.GetRoomByFloor)   // 根据楼层获取房间
			rooms.GET("/search/type", roomHandler.SearchRoomsByType) // 按房型搜索
			rooms.GET("/:id", roomHandler.GetRoomByID)               // 获取房间详情

			// 需要认证的房间管理路由（管理员）
			roomsAuth := rooms.Group("")
			roomsAuth.Use(middleware.AuthMiddleware(), middleware.AdminMiddleware())
			{
				roomsAuth.POST("", roomHandler.CreateRoom)             // 创建房间
				roomsAuth.POST("/batch", roomHandler.BatchCreateRooms) // 批量创建房间
				roomsAuth.POST("/:id", roomHandler.UpdateRoom)         // 更新房间
				roomsAuth.POST("/:id/delete", roomHandler.DeleteRoom)  // 删除房间
			}
		}
		roomCategories := api.Group("/room-categories")
		{
			roomCategories.GET("", roomCategoryHandler.ListRoomCategories)
			roomCategories.GET("/:id", roomCategoryHandler.GetRoomCategoryByID)
		}
		// 设置路由（公开查询）
		settings := api.Group("/settings")
		{
			settings.GET("/public", hotelSettingsHandler.GetPublicSettings)
		}
		// 活动横幅路由（公开查询）
		banners := api.Group("/banners")
		{
			banners.GET("/active", bannerHandler.GetActiveBanners) // 获取激活的活动横幅（前端展示用）
			banners.GET("/:id", bannerHandler.GetBannerPublicByID) // 获取活动横幅详情（前端展示用）
		}
		// 公告路由（公开查询）
		notices := api.Group("/notices")
		{
			notices.GET("/active", noticeHandler.GetActiveNotices) // 获取激活的公告（前端展示用）
		}
		// 推荐路由（公开查询，带 token 时自动个性化）
		recommendations := api.Group("/recommendations")
		{
			recommendations.GET("/rooms", recommendationHandler.GetRoomRecommendations)
		}
		// 日志路由
		logs := api.Group("/logs")
		{
			logs.POST("/report", logHandler.Report) // 上报日志
		}

		// 文件上传路由（需要认证，但不需要管理员权限）
		upload := api.Group("/upload")
		upload.Use(middleware.AuthMiddleware())
		{
			upload.POST("/image", cosHandler.UploadImage) // 通用图片上传接口
		}

		// 需要认证的路由
		authorized := api.Group("")
		authorized.Use(middleware.AuthMiddleware())
		{
			// 用户路由
			users := authorized.Group("/users")
			{
				users.GET("/profile", userHandler.GetProfile)       // 获取个人信息
				users.POST("/profile", userHandler.UpdateProfile)   // 更新个人信息
				users.POST("/password", userHandler.ChangePassword) // 修改密码
			}

			// 预订路由
			bookings := authorized.Group("/bookings")
			{
				bookings.POST("", bookingHandler.CreateBooking)            // 创建预订
				bookings.GET("/my", bookingHandler.GetMyBookings)          // 我的预订列表
				bookings.GET("/:id", bookingHandler.GetBookingByID)        // 获取预订详情
				bookings.POST("/:id/cancel", bookingHandler.CancelBooking) // 取消预订
			}

			recommendationsAuth := authorized.Group("/recommendations")
			{
				recommendationsAuth.POST("/behavior", recommendationHandler.TrackBehavior)
			}

			// 管理员路由
			admin := authorized.Group("/admin")
			admin.Use(middleware.AdminMiddleware())
			{
				admin.GET("/hotels", hotelHandler.ListHotels)
				admin.POST("/hotels", hotelHandler.CreateHotel)
				admin.GET("/hotels/:id", hotelHandler.GetHotelByID)
				admin.POST("/hotels/:id", hotelHandler.UpdateHotel)
				admin.POST("/hotels/:id/delete", hotelHandler.DeleteHotel)

				admin.GET("/settings", hotelSettingsHandler.GetAdminSettings)
				admin.POST("/settings/save", hotelSettingsHandler.SaveSettings)

				// 用户管理
				admin.GET("/users", userHandler.ListUsers)
				admin.GET("/users/:id", userHandler.GetUserByID)
				admin.POST("/users/:id", userHandler.UpdateUser)
				admin.POST("/users/user", userHandler.AddUser)
				admin.POST("/users/batch", userHandler.DeleteUsers)
				// 预订管理
				admin.GET("/bookings", bookingHandler.ListAllBookings)
				admin.GET("/bookings/search", bookingHandler.SearchBookingsByGuestInfo) // 通过客人信息搜索预订
				admin.POST("/bookings/:id/confirm", bookingHandler.ConfirmBooking)
				admin.POST("/bookings/:id/checkin", bookingHandler.CheckIn)
				admin.POST("/bookings/:id/checkout", bookingHandler.CheckOut)
				admin.GET("/bookings/room", bookingHandler.GetBookingsByRoomNumberAndStatus) // 根据房间号和状态获取预订列表

				// 工单与客房服务管理
				admin.POST("/work-orders/repair", workOrderHandler.CreateRepairRequest)
				admin.POST("/work-orders/repair/:id/complete", workOrderHandler.CompleteRepair)
				admin.GET("/work-orders/repairs", workOrderHandler.ListMaintenance)
				admin.POST("/work-orders/cleaning", workOrderHandler.CreateCleaningTask)
				admin.POST("/work-orders/cleaning/:id/assign", workOrderHandler.AssignStaff)
				admin.POST("/work-orders/cleaning/:id/complete", workOrderHandler.CompleteCleaning)
				admin.GET("/work-orders/cleanings", workOrderHandler.ListHousekeeping)

				// 房态库存管理
				admin.POST("/inventory/init", inventoryHandler.InitInventory)
				admin.POST("/inventory/update", inventoryHandler.UpdateInventory)
				admin.GET("/inventory/grid", inventoryHandler.GetInventoryGrid)

				// 定价规则管理
				admin.GET("/pricing/rules", pricingHandler.ListRules)
				admin.POST("/pricing/rules", pricingHandler.CreateRule)

				// 日志管理
				admin.GET("/logs", logHandler.GetLogs) // 获取日志列表
				// 设施管理
				admin.GET("/facilities", facilityHandler.FindAllFacilities)                  // 查询所有设施
				admin.POST("/facilities", facilityHandler.CreateFacility)                    // 创建设施
				admin.POST("/facilities/batch", facilityHandler.BatchUpdateFacilities)       // 批量更新设施位置
				admin.GET("/facilities/floor/:floor", facilityHandler.FindFacilitiesByFloor) // 按楼层查询设施
				admin.GET("/facilities/:id", facilityHandler.FindFacilityByID)               // 根据ID查找设施
				admin.POST("/facilities/:id", facilityHandler.UpdateFacility)                // 更新设施
				admin.POST("/facilities/:id/delete", facilityHandler.DeleteFacility)         // 删除设施

				// 房型分类管理
				admin.GET("/room-categories", roomCategoryHandler.ListRoomCategories)
				admin.GET("/room-categories/:id", roomCategoryHandler.GetRoomCategoryByID)
				admin.POST("/room-categories", roomCategoryHandler.CreateRoomCategory)
				admin.POST("/room-categories/:id", roomCategoryHandler.UpdateRoomCategory)
				admin.POST("/room-categories/:id/delete", roomCategoryHandler.DeleteRoomCategory)

				// 活动横幅管理
				admin.GET("/banners", bannerHandler.GetAllBanners)            // 获取所有活动横幅
				admin.POST("/banners", bannerHandler.CreateBanner)            // 创建活动横幅
				admin.GET("/banners/:id", bannerHandler.GetBannerByID)        // 获取活动横幅详情
				admin.POST("/banners/:id", bannerHandler.UpdateBanner)        // 更新活动横幅
				admin.POST("/banners/:id/delete", bannerHandler.DeleteBanner) // 删除活动横幅

				// 公告管理
				admin.GET("/notices", noticeHandler.GetAllNotices)            // 获取所有公告
				admin.POST("/notices", noticeHandler.CreateNotice)            // 创建公告
				admin.GET("/notices/:id", noticeHandler.GetNoticeByID)        // 获取公告详情
				admin.POST("/notices/:id", noticeHandler.UpdateNotice)        // 更新公告
				admin.POST("/notices/:id/delete", noticeHandler.DeleteNotice) // 删除公告
			}
		}
	}

	setupFrontendStaticRoutes(r, frontendDistDir)
}

func shouldAutoUpdateSwagger(serverMode string) bool {
	v := strings.TrimSpace(strings.ToLower(os.Getenv("SWAGGER_AUTO_UPDATE")))
	if v == "" {
		return strings.TrimSpace(strings.ToLower(serverMode)) == "debug"
	}
	return v == "1" || v == "true" || v == "yes" || v == "on"
}

func resolveBackendDir() string {
	cwd, err := os.Getwd()
	if err != nil {
		return "."
	}

	if _, err := os.Stat(filepath.Join(cwd, "cmd", "server", "main.go")); err == nil {
		return cwd
	}
	if _, err := os.Stat(filepath.Join(cwd, "backend", "cmd", "server", "main.go")); err == nil {
		return filepath.Join(cwd, "backend")
	}

	return cwd
}

func resolveFrontendDistDir(backendDir string) string {
	candidates := []string{
		filepath.Join(backendDir, "..", "frontend", "dist"),
		filepath.Join(backendDir, "frontend", "dist"),
		filepath.Join(".", "frontend", "dist"),
	}

	for _, candidate := range candidates {
		indexPath := filepath.Join(candidate, "index.html")
		if info, err := os.Stat(indexPath); err == nil && !info.IsDir() {
			return candidate
		}
	}

	return ""
}

func setupFrontendStaticRoutes(r *gin.Engine, frontendDistDir string) {
	if strings.TrimSpace(frontendDistDir) == "" {
		log.Println("⚠️  未找到 frontend/dist，跳过前端静态资源托管")
		return
	}

	log.Printf("🌐 托管前端静态资源目录: %s", frontendDistDir)

	r.NoRoute(func(c *gin.Context) {
		requestPath := c.Request.URL.Path

		if strings.HasPrefix(requestPath, "/api/") || strings.HasPrefix(requestPath, "/swagger") || requestPath == "/swagger.json" {
			c.String(http.StatusNotFound, "Cannot %s %s", c.Request.Method, requestPath)
			return
		}

		relativePath := strings.TrimPrefix(filepath.Clean(requestPath), string(filepath.Separator))
		if relativePath == "." {
			relativePath = ""
		}

		serveIfExists := func(paths ...string) bool {
			for _, p := range paths {
				fullPath := filepath.Join(frontendDistDir, filepath.FromSlash(p))
				if !strings.HasPrefix(fullPath, filepath.Clean(frontendDistDir)) {
					continue
				}

				info, err := os.Stat(fullPath)
				if err == nil && !info.IsDir() {
					c.File(fullPath)
					return true
				}
			}
			return false
		}

		if relativePath == "" {
			if serveIfExists("index.html") {
				return
			}
		}

		if relativePath != "" {
			if serveIfExists(relativePath) {
				return
			}

			if filepath.Ext(relativePath) == "" && serveIfExists(filepath.ToSlash(filepath.Join(relativePath, "index.html")), "index.html") {
				return
			}
		}

		c.String(http.StatusNotFound, "Cannot %s %s", c.Request.Method, requestPath)
	})
}

func runSwagInit(backendDir string) error {
	cmd := exec.Command("swag", "init", "-g", "cmd/server/main.go", "-o", "docs")
	cmd.Dir = backendDir

	out, err := cmd.CombinedOutput()
	if err != nil {
		trimmed := strings.TrimSpace(string(out))
		if trimmed == "" {
			return err
		}
		return fmt.Errorf("%w: %s", err, trimmed)
	}

	return nil
}