# backend/docs/swagger.yaml
definitions:
  errors.ErrorInfo:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      message:
        example: 请求参数错误
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorInfo'
      success:
        example: false
        type: boolean
    type: object
  models.Banner:
    properties:
      created_at:
        description: 创建时间
        type: string
      effective_active:
        type: boolean
      end_time:
        description: 活动结束时间（可为空）
        type: string
      id:
        description: 主键（自增ID）
        type: integer
      image_url:
        description: 图片URL
        type: string
      in_time_window:
        type: boolean
      link_url:
        description: 跳转链接（可为空）
        type: string
      manual_active:
        type: boolean
      sort:
        description: 排序，数字越小越靠前
        type: integer
      start_time:
        description: 活动开始时间（可为空）
        type: string
      status:
        description: 状态：active, inactive
        type: string
      subtitle:
        description: 副标题（可为空）
        type: string
      title:
        description: 标题
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
host: 127.0.0.1:18080
info:
  contact:
    email: support@gohotel.com
    name: API Support
  description: 这是一个酒店预订管理系统的 RESTful API 服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 酒店管理系统 API
  version: "1.0"
paths:
  /api/admin/banners:
    get:
      consumes:
      - application/json
      description: 获取所有活动横幅，支持分页
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页条数，默认10
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取所有活动横幅
      tags:
      - 活动管理
    post:
      consumes:
      - multipart/form-data
      description: 创建新的活动横幅，使用临时图片URL
      parameters:
      - description: 活动横幅标题
        in: formData
        name: title
        required: true
        type: string
      - description: 活动横幅副标题
        in: formData
        name: subtitle
        type: string
      - description: 临时图片URL（通过通用上传接口获取）
        in: formData
        name: temp_url
        required: true
        type: string
      - description: 点击跳转链接
        in: formData
        name: link_url
        type: string
      - description: 展示顺序
        in: formData
        name: sort
        type: integer
      - description: 活动开始时间
        in: formData
        name: start_time
        type: string
      - description: 活动结束时间
        in: formData
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 创建活动横幅
      tags:
      - 活动管理
  /api/admin/banners/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取活动横幅详情
      parameters:
      - description: 活动横幅ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 根据ID获取活动横幅
      tags:
      - 活动管理
    post:
      consumes:
      - multipart/form-data
      description: 更新活动横幅信息，使用临时图片URL
      parameters:
      - description: 活动横幅ID
        in: path
        name: id
        required: true
        type: string
      - description: 活动横幅标题
        in: formData
        name: title
        type: string
      - description: 活动横幅副标题
        in: formData
        name: subtitle
        type: string
      - description: 临时图片URL（通过通用上传接口获取）
        in: formData
        name: temp_url
        type: string
      - description: 点击跳转链接
        in: formData
        name: link_url
        type: string
      - description: 展示顺序
        in: formData
        name: sort
        type: integer
      - description: 活动开始时间
        in: formData
        name: start_time
        type: string
      - description: 活动结束时间
        in: formData
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 更新活动横幅
      tags:
      - 活动管理
  /api/admin/banners/{id}/delete:
    post:
      consumes:
      - application/json
      description: 删除指定ID的活动横幅
      parameters:
      - description: 活动横幅ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 删除活动横幅
      tags:
      - 活动管理
  /api/admin/bookings:
    get:
      consumes:
      - application/json
      description: 管理员获取所有预订列表，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Booking'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取所有预订（管理员）
      tags:
      - 管理员
  /api/admin/bookings/{id}/checkin:
    post:
      consumes:
      - application/json
      description: 管理员为已确认的预订办理入住
      parameters:
      - description: 预订 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 办理入住（管理员）
      tags:
      - 管理员
  /api/admin/bookings/{id}/checkout:
    post:
      consumes:
      - application/json
      description: 管理员为入住中的预订办理退房
      parameters:
      - description: 预订 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 办理退房（管理员）
      tags:
      - 管理员
  /api/admin/bookings/{id}/confirm:
    post:
      consumes:
      - application/json
      description: 管理员确认待处理的预订
      parameters:
      - description: 预订 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 确认预订（管理员）
      tags:
      - 管理员
  /api/admin/bookings/room:
    get:
      consumes:
      - application/json
      description: 管理员根据房间号和状态获取预订列表
      parameters:
      - description: 房间号
        in: query
        name: room_number
        required: true
        type: string
      - description: 预订状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"data\": [...], \"count\": number}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 根据房间号和状态获取预订列表
      tags:
      - 管理员
  /api/admin/bookings/search:
    get:
      consumes:
      - application/json
      description: 根据客人姓名、手机号和状态搜索预订记录
      parameters:
      - description: 客人姓名
        in: query
        name: guest_name
        type: string
      - description: 客人手机号
        in: query
        name: guest_phone
        type: string
      - description: 预订状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"data\": [...], \"count\": number}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{\"error\": string}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 通过客人信息搜索预订
      tags:
      - 管理员
  /api/admin/facilities:
    get:
      consumes:
      - application/json
      description: 管理员查询所有设施（分页）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Facility'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 查询所有设施（管理员）
      tags:
      - 管理员
    post:
      consumes:
      - application/json
      description: 管理员创建设施
      parameters:
      - description: 设施信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateFacilityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Facility'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 创建设施（管理员）
      tags:
      - 管理员
...
  /api/admin/facilities/{id}/delete:
    post:
      consumes:
      - application/json
      description: 管理员删除设施
      parameters:
      - description: 设施 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 删除设施（管理员）
      tags:
      - 管理员
...
  /api/admin/facilities/floor/{floor}:
    get:
      consumes:
      - application/json
      description: 管理员获取指定楼层的所有设施
      parameters:
      - description: 楼层
        in: path
        name: floor
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Facility'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 按楼层查询设施（管理员）
      tags:
      - 管理员
...
securityDefinitions:
  Bearer:
    description: 输入 "Bearer {token}" 格式的 JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"